requirements.txt
streamlit
python-dotenv
pdf2image
google-generativeai


